+++ fractal_server/app/schemas/applyworkflow.py
@@ -6,17 +6,58 @@ from typing import Union
 
 from pydantic import BaseModel
 from pydantic import validator
+from pydantic.types import StrictStr
 
 from ._validators import valstr
-from .dataset import DatasetRead
 from .project import ProjectRead
-from .workflow import WorkflowRead
+
 
 __all__ = (
     "_ApplyWorkflowBase",
     "ApplyWorkflowCreate",
     "ApplyWorkflowRead",
 )
+
+
+class TaskDump(BaseModel):
+    id: int
+    source: str
+    name: str
+    command: str
+    input_type: str
+    output_type: str
+    owner: Optional[str]
+    version: Optional[str]
+
+
+class WorkflowTaskDump(BaseModel):
+    id: int
+    order: Optional[int]
+    workflow_id: int
+    task_id: int
+    task: TaskDump
+
+
+class WorkflowDump(BaseModel):
+    id: int
+    name: str
+    project_id: int
+    task_list: list[WorkflowTaskDump]
+
+
+class ResourceDump(BaseModel):
+    id: int
+    path: str
+    dataset_id: int
+
+
+class DatasetDump(BaseModel):
+    id: int
+    name: str
+    type: Optional[str]
+    read_only: bool
+    resource_list: list[ResourceDump]
+    project_id: int
 
 
 class JobStatusType(str, Enum):
@@ -62,10 +103,12 @@ class ApplyWorkflowCreate(_ApplyWorkflowBase):
     Attributes:
         first_task_index:
         last_task_index:
+        slurm_account:
     """
 
     first_task_index: Optional[int] = None
     last_task_index: Optional[int] = None
+    slurm_account: Optional[StrictStr] = None
 
     # Validators
     _worker_init = validator("worker_init", allow_reuse=True)(
@@ -114,6 +157,7 @@ class ApplyWorkflowRead(_ApplyWorkflowBase):
         project_id:
         project_dump:
         user_email:
+        slurm_account:
         workflow_id:
         workflow_dump:
         input_dataset_id:
@@ -134,12 +178,13 @@ class ApplyWorkflowRead(_ApplyWorkflowBase):
     project_id: Optional[int]
     project_dump: Optional[Union[ProjectRead, dict[str, Any]]]
     user_email: str
+    slurm_account: Optional[str]
     workflow_id: Optional[int]
-    workflow_dump: Optional[WorkflowRead]
+    workflow_dump: Optional[WorkflowDump]
     input_dataset_id: Optional[int]
-    input_dataset_dump: Optional[DatasetRead]
+    input_dataset_dump: Optional[DatasetDump]
     output_dataset_id: Optional[int]
-    output_dataset_dump: Optional[DatasetRead]
+    output_dataset_dump: Optional[DatasetDump]
     start_timestamp: datetime
     end_timestamp: Optional[datetime]
     status: str